requirements.txt
wait-for-it==2.2.0
pytest==6.2.5
pre-commit==2.15.0
python-dotenv==1.0.0
pydantic==1.10.7
redis==4.5.5
orjson==3.8.10
httptools==0.5.0
gunicorn==20.1.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
backoff==2.2.1
flask==2.3.2
flask-marshmallow==0.15.0
flask_sqlalchemy==3.0.3
flask-jwt-extended==4.4.4
requests==2.30.0
gevent==22.10.2
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
passlib==1.7.4
flask_jwt_extended==4.4.4
marshmallow-sqlalchemy==0.29.0
flask-migrate==4.0.4
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0